refactor: logs
